requirements.txt
pytest==7.1.1
requests==2.27.1

PyYAML~=6.0